+++ src/main/java/roomescape/dto/PaymentRequest.java
@@ -1,9 +1,11 @@
 package roomescape.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentRequest(
         long reservationId,
         String paymentKey,
         String orderId,
-        long amount
+        BigDecimal amount
 ) {
 }
