+++ src/main/java/roomescape/dto/PaymentResponse.java
@@ -1,9 +1,11 @@
 package roomescape.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentResponse(
         long id,
         String paymentKey,
         String orderId,
-        long totalAmount
+        BigDecimal totalAmount
 ) {
 }
