+++ src/main/java/roomescape/dto/service/TossPaymentResponse.java
@@ -1,11 +1,12 @@
 package roomescape.dto.service;
 
+import java.math.BigDecimal;
 import roomescape.domain.Payment;
 
 public record TossPaymentResponse(
         String paymentKey,
         String orderId,
-        Long totalAmount
+        BigDecimal totalAmount
 ) {
     public Payment toPayment() {
         return new Payment(null, paymentKey, orderId, totalAmount);
